refactor:TripPage 훅으로 정리, 불필요한 코드 제거
